requirements.txt
aniso8601==7.0.0
anyio==3.5.0
APScheduler==3.6.3
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
asyncio==3.4.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
fastapi==0.75.0
fyers-apiv2==2.0.5
getmac==0.8.3
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.13.0
httpie==3.1.0
httptools==0.2.0
idna==2.10
ipykernel==6.9.2
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
lxml==4.8.0
lz4==3.1.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.4
nsepy==0.8
numpy==1.22.2
orjson==3.6.7
outcome==1.1.0
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
promise==2.3
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pyOpenSSL==22.0.0
PyQt6==6.2.3
PyQt6-Qt6==6.2.4
PyQt6-sip==13.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==5.4.1
pyzmq==22.3.0
razorpay==1.3.0
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
Rx==1.6.1
selenium==4.1.2
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.17.1
text-unidecode==1.3
tk==0.1.0
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
truedata-ws==4.2.4
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.8.1
wcwidth==0.2.5
websocket-client==1.2.1
websockets==10.2
wsproto==1.1.0
